requirements.txt
streamlit
sqlalchemy
alembic
pydantic
python-dotenv
pytest
torch
torchvision
pillow
numpy
scikit-learn
faiss-cpu
matplotlib
